refactor: address review feedback — remove backward compat shim, use Wire.* pattern
- Delete src/wire.ts; update all consumers (src/ and tests/) to import
  directly from shared/wire.js
- Use `import * as Wire` in task.ts and task-manager.ts; replace
  WireTask alias with Wire.Task, BlockerInfo with Wire.BlockerInfo
- Remove re-exports from admin-ws.ts
- Replace inline import in tests/admin-ws.test.ts with named import
- Fix frontend/src/types.ts: import TaskStatus before using in TaskRow,
  remove inline import
