+++ src/foreman/admin-ws.ts
@@ -3,7 +3,6 @@ import { WebSocketServer } from "ws";
 import type { WebSocket as WsSocket } from "ws";
 import type { Task, Worker, LogEntry, AdminSnapshot, AdminMessage } from "../../shared/wire.js";
 
-export type { Task, Worker, LogEntry, AdminSnapshot, AdminMessage };
 
 export interface AdminWss {
   broadcastSnapshot(snapshot: AdminSnapshot): void;
