--- src/wire.ts
@@ -1,3 +1,0 @@
-// Re-export all wire types from shared/wire.ts.
-// Consumers can keep importing from here; new code should prefer shared/wire.ts directly.
-export * from "../shared/wire.js";
