+++ tests/foreman.reconcile.test.ts
@@ -7,7 +7,7 @@ import { Task } from "../src/foreman/models/task.js";
 import { setupInMemoryTasks } from "./helpers/task.js";
 import { loadDefaultConfig } from "../src/config.js";
 const defaultCfg = await loadDefaultConfig();
-import * as Wire from "../src/wire.js";
+import * as Wire from "../shared/wire.js";
 
 const TASK_LABEL = "brunel:ready";
