+++ tests/foreman.registry.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { Worker } from "../src/foreman/models/worker.js";
-import * as Wire from "../src/wire.js";
+import * as Wire from "../shared/wire.js";
 
 function fakeWs() {
   return { send: vi.fn(), close: vi.fn(), readyState: 1 } as any;
