+++ tests/foreman.timestamps.test.ts
@@ -12,7 +12,7 @@ import { Task } from "../src/foreman/models/task.js";
 import { setupInMemoryTasks } from "./helpers/task.js";
 import { loadDefaultConfig } from "../src/config.js";
 const defaultCfg = await loadDefaultConfig();
-import * as Wire from "../src/wire.js";
+import * as Wire from "../shared/wire.js";
 import { waitUntil } from "./helpers.js";
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
