+++ tests/foreman.websocket.test.ts
@@ -10,7 +10,7 @@ import { Task } from "../src/foreman/models/task.js";
 import { setupInMemoryTasks } from "./helpers/task.js";
 import { loadDefaultConfig } from "../src/config.js";
 const defaultCfg = await loadDefaultConfig();
-import * as Wire from "../src/wire.js";
+import * as Wire from "../shared/wire.js";
 import { ForemanMessage } from "../src/foreman/models/foreman-message.js";
 import { waitUntil } from "./helpers.js";
